breaking change! warp size for David Andersen's kernels was reduced from 128 threads to 32 threads (to match CUDA definition of a warp). This changes kernel launch configurations! But memory allocation granularity improves...

+++ cpu-miner.c
@@ -694,7 +694,7 @@ static void stratum_gen_work(struct stratum_ctx *sctx, struct work *work)
 	memcpy(work->xnonce2, sctx->job.xnonce2, sctx->xnonce2_size);
 
 	/* Generate merkle root */
-	sha256d(merkle_root, sctx->job.coinbase, sctx->job.coinbase_size);
+	sha256d(merkle_root, sctx->job.coinbase, (int)sctx->job.coinbase_size);
 	for (i = 0; i < sctx->job.merkle_count; i++) {
 		memcpy(merkle_root + 32, sctx->job.merkle[i], 32);
 		sha256d(merkle_root, merkle_root, 64);
